Extract Layer model specs to shared examples
